Wiscasset Music Database Application code
Fixed the "Add music product" tab to create a new product_id that is used throughout the database and as a primary key in the music_product table.  
Added an update function to the music_product catalog to update prices, add data to blank attribute fields etc.  Fixed a bug in the Replenish Inventory tab.